Moved all buff logic to ObjAIBase
