+++ GameServerLib/Logic/GameObjects/Unit.cs
@@ -83,9 +83,6 @@ namespace LeagueSandbox.GameServer.Logic.GameObjects
         private float _statUpdateTimer;
         private uint _autoAttackProjId;
         public MoveOrder MoveOrder { get; set; }
-        private Buff[] AppliedBuffs { get; }
-
-        public List<BuffGameScriptController> BuffGameScriptControllers { get; private set; }
 
         /// <summary>
         /// Unit we want to attack as soon as in range
@@ -111,8 +108,6 @@ namespace LeagueSandbox.GameServer.Logic.GameObjects
         protected Logger _logger = Program.ResolveDependency<Logger>();
 
         public int KillDeathCounter { get; protected set; }
-        private object _buffsLock = new object();
-        private Dictionary<string, Buff> _buffs = new Dictionary<string, Buff>();
 
         private float _timerUpdate;
 
@@ -131,8 +126,6 @@ namespace LeagueSandbox.GameServer.Logic.GameObjects
         ) : base(x, y, collisionRadius, visionRadius, netId)
 
         {
-            BuffGameScriptControllers = new List<BuffGameScriptController>();
-            AppliedBuffs = new Buff[256];
             Stats = stats;
             Model = model;
             CharData = _game.Config.ContentManager.GetCharData(Model);
@@ -209,42 +202,7 @@ namespace LeagueSandbox.GameServer.Logic.GameObjects
         {
             Stats.RemoveModifier(statModifier);
         }
-
-        public BuffGameScriptController AddBuffGameScript(String buffNamespace, String buffClass, Spell ownerSpell, float removeAfter = -1f, bool isUnique = false)
-        {
-            if (isUnique)
-            {
-                RemoveBuffGameScriptsWithName(buffNamespace, buffClass);
-            }
-
-            BuffGameScriptController buffController = 
-                new BuffGameScriptController(this, buffNamespace, buffClass, ownerSpell, duration: removeAfter);
-            BuffGameScriptControllers.Add(buffController);
-            buffController.ActivateBuff();
-
-            return buffController;
-        }
-        public void RemoveBuffGameScript(BuffGameScriptController buffController)
-        {
-            buffController.DeactivateBuff();
-            BuffGameScriptControllers.Remove(buffController);
-        }
-        public bool HasBuffGameScriptActive(String buffNamespace, String buffClass)
-        {
-            foreach (BuffGameScriptController b in BuffGameScriptControllers)
-            {
-                if (b.IsBuffSame(buffNamespace, buffClass)) return true;
-            }
-            return false;
-        }
-        public void RemoveBuffGameScriptsWithName(String buffNamespace, String buffClass)
-        {
-            foreach (BuffGameScriptController b in BuffGameScriptControllers)
-            {
-                if (b.IsBuffSame(buffNamespace, buffClass)) b.DeactivateBuff();
-            }
-            BuffGameScriptControllers.RemoveAll((b) => b.NeedsRemoved());
-        }
+        
         public void StopMovement()
         {
             this.SetWaypoints(new List<Vector2> { this.GetPosition(), this.GetPosition() });
@@ -263,8 +221,6 @@ namespace LeagueSandbox.GameServer.Logic.GameObjects
                 cc.Update(diff);
             }
             crowdControlList.RemoveAll((cc)=>cc.IsDead());
-
-            BuffGameScriptControllers.RemoveAll((b) => b.NeedsRemoved());
 
             var onUpdate = _scriptEngine.GetStaticMethod<Action<Unit, double>>(Model, "Passive", "OnUpdate");
             onUpdate?.Invoke(this, diff);
@@ -407,23 +363,6 @@ namespace LeagueSandbox.GameServer.Logic.GameObjects
         public override float getMoveSpeed()
         {
             return Stats.MoveSpeed.Total;
-        }
-
-        public Dictionary<string, Buff> GetBuffs()
-        {
-            var toReturn = new Dictionary<string, Buff>();
-            lock (_buffsLock)
-            {
-                foreach (var buff in _buffs)
-                    toReturn.Add(buff.Key, buff.Value);
-
-                return toReturn;
-            }
-        }
-
-        public int GetBuffsCount()
-        {
-            return _buffs.Count;
         }
 
         public override void onCollision(GameObject collider)
@@ -596,48 +535,9 @@ namespace LeagueSandbox.GameServer.Logic.GameObjects
             }
         }
 
-        public void AddBuff(Buff b)
-        {
-            lock (_buffsLock)
-            {
-                if (!_buffs.ContainsKey(b.Name))
-                {
-                    _buffs.Add(b.Name, b);
-                }
-                else
-                {
-                    _buffs[b.Name].TimeElapsed = 0; // if buff already exists, just restart its timer
-                }
-            }
-        }
-
-        public void RemoveBuff(Buff b)
-        {
-            //TODO add every stat
-            RemoveBuff(b.Name);
-            RemoveBuffSlot(b);
-        }
-
-        public void RemoveBuff(string b)
-        {
-            lock (_buffsLock)
-                _buffs.Remove(b);
-        }
-
         public virtual bool isInDistress()
         {
             return false; //return DistressCause;
-        }
-
-        //todo: use statmods
-        public Buff GetBuff(string name)
-        {
-            lock (_buffsLock)
-            {
-                if (_buffs.ContainsKey(name))
-                    return _buffs[name];
-                return null;
-            }
         }
 
         public void SetTargetUnit(Unit target)
@@ -750,31 +650,6 @@ namespace LeagueSandbox.GameServer.Logic.GameObjects
             }
 
             return ClassifyUnit.Default;
-        }
-
-        public byte GetNewBuffSlot(Buff b)
-        {
-            byte slot = GetBuffSlot();
-            AppliedBuffs[slot] = b;
-            return slot;
-        }
-
-        public void RemoveBuffSlot(Buff b)
-        {
-            byte slot = GetBuffSlot(b);
-            AppliedBuffs[slot] = null;
-        }
-
-        private byte GetBuffSlot(Buff buffToLookFor = null)
-        {
-            for (byte i = 1; i < AppliedBuffs.Length; i++) // Find the first open slot or the slot corresponding to buff
-            {
-                if (AppliedBuffs[i] == buffToLookFor)
-                {
-                    return i;
-                }
-            }
-            throw new Exception("No slot found with requested value"); // If no open slot or no corresponding slot
         }
     }
